backend/services/document_ai_service.py
"""Document AI service for resume parsing and document analysis."""

import logging
from typing import Dict, Any, List, Optional
import base64
import mimetypes

from google.cloud import documentai

from core.config import settings

logger = logging.getLogger(__name__)


class DocumentAIService:
    """Service for Google Document AI operations."""
    
    def __init__(self):
        self.project_id = settings.GOOGLE_CLOUD_PROJECT
        self.location = settings.VERTEX_AI_LOCATION  # Using same location as Vertex AI
        self.processor_id = None  # Would be configured separately
        self.client = None
    
    def _initialize_client(self):
        """Initialize Document AI client."""
        try:
            self.client = documentai.DocumentProcessorServiceClient()
            logger.info("Document AI service initialized")
        except Exception as e:
            logger.error(f"Failed to initialize Document AI service: {e}")
            raise
    
    async def parse_resume(self, file_path_or_content: str) -> Dict[str, Any]:
        """Parse resume using Document AI."""
        try:
            if self.client is None:
                self._initialize_client()
            
            # For now, return mock parsed data since Document AI setup requires
            # processor configuration
            return await self._mock_resume_parsing(file_path_or_content)
            
        except Exception as e:
            logger.error(f"Resume parsing failed: {e}")
            return {"error": str(e)}
    
    async def _mock_resume_parsing(self, file_input: str) -> Dict[str, Any]:
        """Mock resume parsing for development."""
        # In a real implementation, this would:
        # 1. Send document to Document AI processor
        # 2. Extract structured information
        # 3. Return parsed data
        
        mock_data = {
            "text": "Sample resume text content...",
            "education": [
                {
                    "degree": "Bachelor of Technology",
                    "field": "Computer Science",
                    "institution": "IIT Delhi",
                    "year": "2023",
                    "grade": "8.5 CGPA"
                }
            ],
            "experience": [
                {
                    "title": "Software Engineering Intern",
                    "company": "Tech Company",
                    "start_date": "2022-06",
                    "end_date": "2022-08",
                    "description": "Worked on web development projects"
                }
            ],
            "skills": [
                "Python", "JavaScript", "React", "Node.js", "SQL",
                "Git", "AWS", "Machine Learning"
            ],
            "contact": {
                "email": "user@example.com",
                "phone": "+91-9876543210",
                "location": "New Delhi, India"
            },
            "certifications": [
                "AWS Certified Developer",
                "Google Cloud Associate"
            ],
            "projects": [
                {
                    "name": "E-commerce Website",
                    "description": "Built a full-stack e-commerce platform",
                    "technologies": ["React", "Node.js", "MongoDB"]
                }
            ],
            "languages": ["English", "Hindi"],
            "parse_confidence": 0.85
        }
        
        return mock_data
    
    async def extract_skills_from_text(self, text: str) -> List[str]:
        """Extract skills from text using Document AI."""
        try:
            # Mock implementation - in production would use Document AI
            # to extract and classify skills from text
            
            common_skills = {
                "programming": ["python", "java", "javascript", "c++", "c#", "go", "rust", "ruby"],
                "web": ["html", "css", "react", "angular", "vue", "nodejs", "express"],
                "data": ["sql", "mongodb", "postgresql", "mysql", "pandas", "numpy"],
                "cloud": ["aws", "azure", "gcp", "docker", "kubernetes"],
                "tools": ["git", "jenkins", "jira", "slack", "figma"]
            }
            
            text_lower = text.lower()
            extracted_skills = []
            
            for category, skills in common_skills.items():
                for skill in skills:
                    if skill in text_lower:
                        extracted_skills.append(skill.title())
            
            return list(set(extracted_skills))
            
        except Exception as e:
            logger.error(f"Skill extraction failed: {e}")
            return []
    
    async def analyze_document_structure(self, file_content: bytes, 
                                       mime_type: str) -> Dict[str, Any]:
        """Analyze document structure and extract key information."""
        try:
            # Mock implementation
            return {
                "document_type": "resume",
                "sections": [
                    {"type": "header", "content": "Contact Information"},
                    {"type": "section", "content": "Education"},
                    {"type": "section", "content": "Work Experience"},
                    {"type": "section", "content": "Skills"},
                    {"type": "section", "content": "Projects"}
                ],
                "confidence": 0.9,
                "language": "en"
            }
            
        except Exception as e:
            logger.error(f"Document analysis failed: {e}")
            return {"error": str(e)}
    
    async def extract_education_info(self, text: str) -> List[Dict[str, Any]]:
        """Extract education information from text."""
        try:
            # Mock implementation - would use NLP to extract education details
            education_info = [
                {
                    "degree": "Bachelor of Technology",
                    "field": "Computer Science Engineering",
                    "institution": "Indian Institute of Technology",
                    "year": "2023",
                    "grade": "8.5 CGPA",
                    "location": "Delhi, India"
                }
            ]
            
            return education_info
            
        except Exception as e:
            logger.error(f"Education extraction failed: {e}")
            return []
    
    async def extract_work_experience(self, text: str) -> List[Dict[str, Any]]:
        """Extract work experience from text."""
        try:
            # Mock implementation
            experience = [
                {
                    "title": "Software Engineering Intern",
                    "company": "Technology Solutions Ltd",
                    "start_date": "June 2022",
                    "end_date": "August 2022",
                    "duration": "3 months",
                    "description": "Developed web applications using React and Node.js",
                    "responsibilities": [
                        "Frontend development with React",
                        "API integration",
                        "Code review and testing"
                    ]
                }
            ]
            
            return experience
            
        except Exception as e:
            logger.error(f"Experience extraction failed: {e}")
            return []
    
    async def validate_document_quality(self, file_content: bytes) -> Dict[str, Any]:
        """Validate document quality for processing."""
        try:
            # Mock implementation - would check image quality, text clarity, etc.
            return {
                "quality_score": 0.85,
                "issues": [],
                "recommendations": [],
                "processable": True
            }
            
        except Exception as e:
            logger.error(f"Document validation failed: {e}")
            return {
                "quality_score": 0.0,
                "issues": [str(e)],
                "recommendations": ["Please upload a clearer document"],
                "processable": False
            }
    
    async def health_check(self) -> bool:
        """Check if Document AI service is healthy."""
        try:
            if self.client is None:
                self._initialize_client()
            # Simple health check - in production would test with sample document
            return True
        except Exception as e:
            logger.error(f"Document AI health check failed: {e}")
            return False