backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud Libraries
google-cloud-aiplatform>=1.39.0
google-cloud-firestore>=2.13.1
google-cloud-bigquery>=3.13.0
google-cloud-documentai>=2.21.0
google-cloud-storage>=2.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Environment & Configuration
python-dotenv==1.0.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.0

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Multi-agent Framework
langchain==0.0.340
langchain-google-vertexai==0.0.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Development
watchfiles==0.21.0